ai-model/requirements.txt
opencv-python
numpy
ultralytics
requests
google-generativeai
pillow
twilio
elevenlabs
python-dotenv
pymongo[srv]